backend/requirements.txt
fastapi
uvicorn[standard]
pinecone-client
sentence-transformers
pandas
python-dotenv